+++ pkg/api/v1alpha2/init.go
@@ -3,5 +3,5 @@ package v1alpha2
 import "github.com/open-platform-model/library/pkg/api"
 
 func init() {
-	api.Register(binding{})
+	api.Register(&binding{})
 }